% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bayes_fosr.R
\name{bayes_fosr}
\alias{bayes_fosr}
\title{Bayesian Function-on-scalar regression}
\usage{
bayes_fosr(formula, data = NULL, est.method = "VB", cov.method = "FPCA",
  ...)
}
\arguments{
\item{formula}{a formula indicating the structure of the proposed model.
Random intercepts are designated using \code{\link{re}}().}

\item{data}{an optional data frame, list or environment containing the
variables in the model. If not found in data, the variables are taken from
environment(formula), typically the environment from which the function is
called.}

\item{est.method}{method used to estimate model parameters. Options are "VB",
"Gibbs", and "GLS" with "VB" as default. Variational Bayes is a fast approximation to
the full posterior and often provides good point estimates, but may be
unreliable for inference. "GLS" doesn't do anything Bayesian -- just fits an
unpenalized GLS estimator for the specified model.}

\item{cov.method}{method used to estimate the residual covariance structure.
Options are "FPCA" and "Wishart", with default "FPCA"}

\item{...}{additional arguments that are passed to individual fitting functions.}
}
\description{
Wrapper function that implements several approaches to Bayesian function-
on-scalar regression. Currently handles real-valued response curves; models
can include subject-level random effects in a multilevel framework. The
residual curve error structure can be estimated using Bayesian FPCA or a
Wishart prior. Model parameters can be estimated using a Gibbs sampler
or variational Bayes.
}
\examples{
\dontrun{
data(DTI)
DTI = subset(DTI, select = c(cca, case, pasat))
DTI = DTI[complete.cases(DTI),]
DTI$gender = factor(sample(c("male","female"), dim(DTI)[1], replace = TRUE))
DTI$status = factor(sample(c("RRMS", "SPMS", "PPMS"), dim(DTI)[1], replace = TRUE))

fosr.dti.default = bayes_fosr(cca ~ pasat, data = DTI)
fosr.dti.vb = bayes_fosr(cca ~ pasat, data = DTI, Kp = 4, Kt = 10)
fosr.dti.gibbs = bayes_fosr(cca ~ pasat, data = DTI, Kp = 4, Kt = 10, est.method = "Gibbs", cov.method = "Wishart")
}
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}
\references{
Goldsmith, J., Kitago, T. (Accepted).
Assessing Systematic Effects of Stroke on Motor Control using Hierarchical
Function-on-Scalar Regression.
Journal of the Royal Statistical Society: Series C.
}

