% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{build_coef_axes}
\alias{build_coef_axes}
\title{Build coordinate axes for coefficient results}
\usage{
build_coef_axes(trm, data_grid)
}
\arguments{
\item{trm}{Smooth term with dim attribute.}

\item{data_grid}{Grid data frame with xm/ym/zm attributes.}
}
\value{
List with x, y, z, xlab, ylab, zlab, xlim, ylim, zlim as appropriate.
}
\description{
Build coordinate axes for coefficient results
}
\keyword{internal}
