% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{coef_safe_range}
\alias{coef_safe_range}
\title{Safely compute range for coef.pffr}
\usage{
coef_safe_range(x)
}
\arguments{
\item{x}{A vector (possibly factor).}
}
\value{
Numeric vector of length 2 with range or c(NA, NA) for factors.
}
\description{
Returns NA range for factors, otherwise numeric range.
}
\keyword{internal}
