% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-robust.R
\name{coefboot.pffr}
\alias{coefboot.pffr}
\title{Simple bootstrap CIs for pffr (deprecated)}
\usage{
coefboot.pffr(
  object,
  n1 = 100,
  n2 = 40,
  n3 = 20,
  B = 100,
  ncpus = getOption("boot.ncpus", 1),
  parallel = c("no", "multicore", "snow"),
  cl = NULL,
  conf = c(0.9, 0.95),
  type = "percent",
  method = c("resample", "residual", "residual.c"),
  showProgress = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted \code{\link{pffr}}-model}

\item{n1}{see \code{\link{coef.pffr}}}

\item{n2}{see \code{\link{coef.pffr}}}

\item{n3}{see \code{\link{coef.pffr}}}

\item{B}{number of bootstrap replicates, defaults to (a measly) 100}

\item{ncpus}{see \code{\link[boot]{boot}}. Defaults to \code{getOption("boot.ncpus", 1L)} (like \code{boot}).}

\item{parallel}{see \code{\link[boot]{boot}}}

\item{cl}{see \code{\link[boot]{boot}}}

\item{conf}{desired levels of bootstrap CIs, defaults to 0.90 and 0.95}

\item{type}{type of bootstrap interval, see \code{\link[boot]{boot.ci}}. Defaults to "percent" for percentile-based CIs.}

\item{method}{either "resample" (default) to resample response trajectories, or "residual" to resample responses as fitted values
plus residual trajectories or "residual.c" to resample responses as fitted values
plus residual trajectories that are centered at zero for each gridpoint.}

\item{showProgress}{TRUE/FALSE}

\item{...}{further arguments handed to \code{\link[boot]{boot}} like e.g. \code{strata} for models with random effects.}
}
\value{
A list with similar structure as the return value of
  \code{\link{coef.pffr}}, containing the original point estimates along
  with bootstrap CIs.
}
\description{
**Deprecated**, as is its successor [pffr_coefboot()]: use the default
intervals of [pffr()] fits instead.
}
\keyword{internal}
