% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.pfr.R
\name{coefficients.pfr}
\alias{coef.pfr}
\alias{coefficients.pfr}
\title{Extract coefficient functions from a fitted pfr-object}
\usage{
\method{coefficients}{pfr}(object, select = NULL, se = TRUE,
  exclude = FALSE, plot = FALSE, ...)

\method{coef}{pfr}(object, select = NULL, se = TRUE, exclude = FALSE,
  plot = FALSE, ...)
}
\arguments{
\item{object}{return object from \code{\link{pfr}}}

\item{select}{which of the smooth terms to extract coefficients from
(integer vector). If \code{NULL}, returns all smooth terms.}

\item{se}{compute standard errors? defaults to TRUE. See \code{\link[mgcv]{plot.gam}}}

\item{exclude}{if \code{TRUE}, excludes reporting of the estimate at coordinates that are
"too far" from data used to fit the model, as determined by
\code{mgcv::plot.mgcv.smooth}, by setting the estimate to \code{NA}.}

\item{plot}{if \code{TRUE}, also plots the coefficient(s)}

\item{...}{further arguments passed on to \pkg{mgcv}'s \code{plot.gam}. Common
arguments include \code{n} and \code{n2} to set the number of coordinates
to estimate for 1-D and 2-D coefficient functions, and \code{seWithMean}
if the standard error should include uncertainty about the overall mean.
See \code{\link[mgcv]{plot.gam}}.}
}
\value{
a list of data.frames containing the evaluation points,
   coefficient function values and optionally their se's for each term in \code{select}.
   If only one term is selected, the one data frame is unlisted.
}
\description{
Extract coefficient functions from a fitted pfr-object
}
\examples{
#TODO: see ?pfr
}
\author{
Fabian Scheipl and Jonathan Gellar, se-computation parts adapted from
   Simon Wood's \code{plot.gam}.
}

