% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{compute_coef_se}
\alias{compute_coef_se}
\title{Compute standard errors for coefficient extraction}
\usage{
compute_coef_se(linear_map, covmat)
}
\arguments{
\item{linear_map}{List returned by build_coef_linear_map().}

\item{covmat}{Covariance matrix.}
}
\value{
Numeric vector of standard errors.
}
\description{
Compute standard errors for coefficient extraction
}
\keyword{internal}
