% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{compute_intercept}
\alias{compute_intercept}
\title{Compute functional intercept}
\usage{
compute_intercept(mu_t, n)
}
\arguments{
\item{mu_t}{Numeric vector (length(yind)) of intercept function values.}

\item{n}{Integer, number of observations.}
}
\value{
Numeric matrix (n x length(mu_t)) of intercept contributions.
}
\description{
Compute the intercept term mu(t) replicated for all observations.
}
\keyword{internal}
