% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{ensure_grid_axis_attributes}
\alias{ensure_grid_axis_attributes}
\title{Ensure evaluation grid has axis attributes used by coef extraction}
\usage{
ensure_grid_axis_attributes(d, trm, is_pcre, pffr_info)
}
\arguments{
\item{d}{Evaluation grid data frame.}

\item{trm}{Smooth term object.}

\item{is_pcre}{Logical, is this a pcre term?}

\item{pffr_info}{List with yind_name.}
}
\value{
Data frame with xm/ym/zm attributes set.
}
\description{
Ensure evaluation grid has axis attributes used by coef extraction
}
\keyword{internal}
