% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{eval_bs}
\alias{eval_bs}
\title{Evaluate B-spline at new points using stored specification}
\usage{
eval_bs(x, spec)
}
\arguments{
\item{x}{Numeric vector of evaluation points.}

\item{spec}{Specification from [make_bs_spec()].}
}
\value{
B-spline basis matrix.
}
\description{
Evaluate B-spline at new points using stored specification
}
\keyword{internal}
