% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_sum2.R
\name{f_sum2}
\alias{f_sum2}
\title{Sum computation 2}
\usage{
f_sum2(y, fixef, mu.q.c, kt, theta)
}
\arguments{
\item{y}{outcome matrix}

\item{fixef}{current estimate of fixed effects}

\item{mu.q.c}{current value of mu.q.c}

\item{kt}{number of basis functions}

\item{theta}{spline basis}
}
\description{
Internal function used compute a sum in FPCA-based covariance updates
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

