% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/f_sum4.R
\name{f_sum4}
\alias{f_sum4}
\title{Sum computation 2}
\usage{
f_sum4(mu.q.c, sig.q.c, mu.q.bpsi, sig.q.bphi, theta, obspts.mat)
}
\arguments{
\item{mu.q.c}{}

\item{sig.q.c}{}

\item{theta}{}

\item{obspts.mat}{}

\item{mu.q.bspi}{}

\item{sig.q.bpsi}{}
}
\description{
Internal function used compute a sum in FPCA-based covariance updates
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

