% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/f_trace.R
\name{f_trace}
\alias{f_trace}
\title{Trace computation}
\usage{
f_trace(Theta_i, Sig_q_Bpsi, Kp, Kt)
}
\arguments{
\item{Theta_i}{}

\item{Sig_q_Bpsi}{}

\item{Kp}{}

\item{Kt}{}
}
\description{
Internal function used compute a trace in FPCA-based covariance updates
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

