% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{finalize_grid_by_var}
\alias{finalize_grid_by_var}
\title{Add by-variable column to grid if needed}
\usage{
finalize_grid_by_var(d, trm)
}
\arguments{
\item{d}{Data frame grid.}

\item{trm}{Smooth term object.}
}
\value{
Modified data frame with by column set to 1 if applicable.
}
\description{
Add by-variable column to grid if needed
}
\keyword{internal}
