% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fpc.R
\name{fpc}
\alias{fpc}
\title{Construct a FPC regression term}
\usage{
fpc(X, ncomp = NULL, pve = 0.99, bs = "ps", k = 40, ...)
}
\arguments{
\item{X}{functional predictors, expressed as an \code{N} by \code{J} matrix,
where \code{N} is the number of columns and \code{J} is the number of
evaluation points. May include missing/sparse functions, which are
indicated by \code{NA} values.}

\item{ncomp}{number of principal components. if \code{NULL}, chosen by \code{pve}}

\item{pve}{proportion of variance explained; used to choose the number of
principal components}

\item{bs}{two letter character string indicating the \code{mgcv}-style basis
to use for pre-smoothing \code{X}}

\item{k}{the dimension of the pre-smoothing basis}

\item{...}{additional options to be passed to \code{\link{lf}}. These include
  \code{argvals}, \code{integration}, and any additional options for the
  pre-smoothing basis (as constructed by \code{mgcv::s}), such as \code{m}.}
}
\value{
The result of a call to \code{\link{lf}}.
}
\description{
Constructs a functional principal component regression (Reiss and Ogden,
2007, 2010) term for inclusion in an \code{mgcv::gam}-formula (or
\code{\link{bam}} or \code{\link{gamm}} or \code{gamm4:::gamm}) as
constructed by \code{\link{pfr}}. Currently only one-dimensional functions
are allowed.
}
\details{
\code{fpc} is a wrapper for \code{\link{lf}}, which defines linear
functional predictors for any type of basis for inclusion in a \code{pfr}
formula. \code{fpc} simply calls \code{lf} with the appropriate options for
the \code{fpc} basis and penalty construction.

The method is the FPCR-R method of Reiss and Ogden (2007). This
  method is also implemented in \code{\link{fpcr}}; here we implement the
  method for inclusion in a \code{pfr} formula.
}
\examples{
data(gasoline)
gasmod <- pfr(octane ~ fpc(NIR, ncomp=30), data=gasoline)
plot(gasmod, rug=FALSE)
est <- coef(gasmod)
}
\author{
Jonathan Gellar \email{JGellar@mathematica-mpr.com}
}
\seealso{
\code{\link{lf}}, \code{\link{smooth.construct.fpc.smooth.spec}}
}

