% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fpca.R
\name{fpca}
\alias{fpca}
\title{Functional Principal Component Analysis}
\usage{
fpca(X, argvals = seq(0, 1, length = ncol(X)), method = "sc",
  random.int = FALSE, nbasis = 10, pve = 0, npc = NULL, var = FALSE,
  simul = FALSE, sim.alpha = 0.95, useSymm = FALSE, makePD = FALSE,
  center = TRUE, cov.est.method = 2)
}
\description{
Functional Principal Component Analysis
}

