% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{gen_random_2d}
\alias{gen_random_2d}
\title{Generate 2D random surface from tensor P-spline prior}
\usage{
gen_random_2d(s_grid, t_grid, k_s, k_t, wiggliness)
}
\arguments{
\item{s_grid}{First dimension evaluation points.}

\item{t_grid}{Second dimension evaluation points.}

\item{k_s}{Basis dimension for s.}

\item{k_t}{Basis dimension for t.}

\item{wiggliness}{Controls smoothness.}
}
\value{
Matrix (length(s_grid) x length(t_grid)), scaled to unit SD.
}
\description{
Samples from tensor product P-spline prior and evaluates on fixed grids.
Returns a matrix scaled to unit SD.
}
\keyword{internal}
