% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{get_default_effect}
\alias{get_default_effect}
\title{Get default effect specification for a term type}
\usage{
get_default_effect(term_type, varname, wrapped = FALSE)
}
\arguments{
\item{term_type}{Character string of term type.}

\item{varname}{Character string of variable name.}
}
\value{
Default effect specification.
}
\description{
Get default effect specification for a term type
}
\keyword{internal}
