% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{get_smooth_label_from_term}
\alias{get_smooth_label_from_term}
\title{Get smooth label from a term string}
\usage{
get_smooth_label_from_term(x)
}
\description{
Get smooth label from a term string
}
\keyword{internal}
