% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getEst.R
\name{getEst}
\alias{getEst}
\title{Extract coefficient function estimate from a fitted fgam model}
\usage{
getEst(mod, data, sm.id = 1, var = FALSE, lincom = NULL)
}
\arguments{
\item{mod}{A fitted \code{fgam} model.}

\item{data}{A data frame containing the locations at which the term will
be evaluated. See Details.}

\item{sm.id}{The index of the smooth term in the model. Only smooth
terms are given an index number. May be supplied as a vector; see
Details.}

\item{var}{Should the pointwise variance also be returned?}

\item{lincom}{Matrix of coefficients to use for a linear combination
   across multiple terms. Should have the same number of columns as
   \code{length(sm.id)}.This feature assumes all bases for the smooth
   terms defined by sm.id are identical. See Details.}
}
\value{
A list, matrix, or vector containing the resulting coefficient
function estimate(s) and/or pointwise variance estimate(s).
}
\description{
This function may be used to extract a functional coefficient from
a fitted \code{fgam} model, at user-specified evaluation points. It works
for \code{lf}, \code{af}, and \code{lf.vd} terms. This function
will be depricated once we write a proper \code{coef.fgam} method.
Can also perform a linear combination of multiple smooth terms,
which is useful for \code{lf.vd} terms with parametric interactions.
}
\details{
The index indicated by \code{sm.id} corresponds to the index
within the \code{mod$smooth} object. There will be one entry in this
list for each smooth term. Use \code{names(mod$smooth)} to see a label
for each smooth term in the model, which should allow the user to
determine the appropriate \code{sm.id}.

The labels produced by \code{names(mod$smooth)} also contains the name
of all variables that are required to be included in the \code{data}
argument. If a "by" matrix (which will have a variable name beginning with
the letter "L") is not supplied, it will be assumed to be 1 (no weights),
which is usually desired. The other variables that are supplied will include
the indices of the appropriate evaluation points. For univariate
smooths (e.g., \code{lf} terms and \code{lf.vd} terms with parametric
interactions), the evaluation points should be supplied as vector columns
in the data frame. For bivariate smooths (e.g., \code{af} terms and fully
nonparametric \code{lf.vd} terms), the evaluation points should be supplied
as matrix-columns in the data frame, with the same dimension for each one.
Note that matrices can be included as columns in the data frame by using
the \code{I()} ("AsIs") function, or by simply assigning them to an existing
dataframe (\code{df$matcolumn <- myMatrix}).

If \code{sm.id} is supplied as a vector with \code{lincom} equal to
\code{NULL}, then multiple smooth terms will be returned. If
\code{lincom} is supplied, its rows will contain the weights for a linear
combination of multiple estimates. e.g., to perform the linear combination
\eqn{\beta_1(t) + T_i \beta_2(t) + T_i^2 \beta_3(t)}, \code{lincom} will be
a matrix of 3 columns: a column of 1's, a column of \eqn{T_i} values, and
a column of \eqn{T_i^2} values. \code{sm.id} will be a vector corresponding
to the indecies of \eqn{\beta_1(t)}, \eqn{\beta_2(t)}, and \eqn{\beta_3(t)}
in the model, and \code{data} will be a data frame with vector columns for
each variable. See examples.
}
\author{
Jonathan E. Gellar <jgellar1@jhu.edu>
}
\seealso{
\code{\link{fgam}}, \code{\link{lf.vd}} for examples.
}

