% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf.vd.R
\name{getL}
\alias{getL}
\title{Get the weight matrix for a linear functional term}
\usage{
getL(tind, integration, n.int = NULL)
}
\description{
Get the weight matrix for a linear functional term
}
\keyword{internal}

