% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{intercept_presets}
\alias{intercept_presets}
\title{Intercept presets for mu(t)}
\usage{
intercept_presets
}
\description{
Named list of preset intercept functions mu(t).
Each function takes argument t and returns a vector.
}
\keyword{internal}
