% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{linear_presets}
\alias{linear_presets}
\title{Linear varying coefficient presets}
\usage{
linear_presets
}
\description{
Named list of preset beta(t) functions for linear varying coefficient terms.
Each function takes argument t and returns a vector.
}
\keyword{internal}
