% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{n_clusters_checked}
\alias{n_clusters_checked}
\title{Number of clusters, requiring at least two}
\usage{
n_clusters_checked(cluster_id)
}
\arguments{
\item{cluster_id}{Cluster membership vector.}
}
\value{
The number of distinct clusters.
}
\description{
The cluster sandwich's G/(G-1) small-sample factor is undefined for a
single cluster.
}
\keyword{internal}
