% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{parse_single_term}
\alias{parse_single_term}
\title{Parse a single term string}
\usage{
parse_single_term(term_str, specials)
}
\arguments{
\item{term_str}{Character string of the term.}

\item{specials}{Character vector of special term types.}
}
\value{
A list with `type`, `varname`, and `call`.
}
\description{
Parse a single term string
}
\keyword{internal}
