% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{pffr.check}
\alias{pffr.check}
\title{Some diagnostics for a fitted pffr model (deprecated)}
\usage{
pffr.check(
  b,
  old.style = FALSE,
  type = c("deviance", "pearson", "response"),
  k.sample = 5000,
  k.rep = 200,
  rep = 0,
  level = 0.9,
  rl.col = 2,
  rep.col = "gray80",
  ...
)
}
\arguments{
\item{b}{a fitted \code{\link{pffr}}-object}

\item{old.style}{If you want old fashioned plots, exactly as in Wood, 2006, set to \code{TRUE}.}

\item{type}{type of residuals, see \code{\link[mgcv]{residuals.gam}}, used in
    all plots.}

\item{k.sample}{Above this k testing uses a random sub-sample of data.}

\item{k.rep}{how many re-shuffles to do to get p-value for k testing.}

\item{rep}{passed to \code{\link[mgcv]{qq.gam}} when \code{old.style} is \code{FALSE}.}

\item{level}{passed to \code{\link[mgcv]{qq.gam}} when \code{old.style} is \code{FALSE}.}

\item{rl.col}{passed to \code{\link[mgcv]{qq.gam}} when \code{old.style} is \code{FALSE}.}

\item{rep.col}{passed to \code{\link[mgcv]{qq.gam}} when \code{old.style} is \code{FALSE}.}

\item{...}{extra graphics parameters to pass to plotting functions.}
}
\value{
None, called for its side effect of producing diagnostic plots.
}
\description{
**Deprecated**

`pffr.check()` was renamed to [pffr_check()] for consistency with the
package naming conventions.
}
\keyword{internal}
