% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_attach_metadata}
\alias{pffr_attach_metadata}
\title{Attach pffr metadata to model}
\usage{
pffr_attach_metadata(m, algorithm, ret)
}
\arguments{
\item{m}{Fitted model.}

\item{algorithm}{Algorithm symbol.}

\item{ret}{pffr metadata list.}
}
\value{
Model with pffr class and metadata.
}
\description{
Attach pffr metadata to model
}
\keyword{internal}
