% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_build_ar_start}
\alias{pffr_build_ar_start}
\title{Build AR.start indicator for AR(1) errors}
\usage{
pffr_build_ar_start(response_name, formula_env, obs_indices)
}
\arguments{
\item{response_name}{Response variable name.}

\item{formula_env}{Formula environment.}

\item{obs_indices}{Observation indices.}
}
\description{
Build AR.start indicator for AR(1) errors
}
\keyword{internal}
