% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_cl2_unavailable}
\alias{pffr_cl2_unavailable}
\title{Why the CL2 sandwich is unavailable for a fit, if it is}
\usage{
pffr_cl2_unavailable(m, algorithm, cluster_id)
}
\arguments{
\item{m}{The fitted gam (or gamm) object.}

\item{algorithm}{Algorithm name.}

\item{cluster_id}{Cluster vector of the fit (one entry per fitted
observation).}
}
\value{
`NULL` if CL2 is available, otherwise a short reason.
}
\description{
Why the CL2 sandwich is unavailable for a fit, if it is
}
\keyword{internal}
