% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_configure_algorithm}
\alias{pffr_configure_algorithm}
\title{Select and configure algorithm}
\usage{
pffr_configure_algorithm(algorithm, ntotal, call, where_specials, use_ar)
}
\arguments{
\item{algorithm}{User-specified algorithm (may be NA).}

\item{ntotal}{Total number of data points.}

\item{call}{The matched call (modified in place for method).}

\item{where_specials}{List of special term indices.}

\item{use_ar}{Whether AR(1) errors requested.}
}
\value{
Symbol for the algorithm.
}
\description{
Select and configure algorithm
}
\keyword{internal}
