% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_expand_variables}
\alias{pffr_expand_variables}
\title{Expand variables to formula environment for smooth/parametric terms}
\usage{
pffr_expand_variables(
  terms,
  where_specials,
  is_sparse,
  nyindex,
  nobs,
  obs_indices,
  eval_env,
  formula_env
)
}
\arguments{
\item{terms}{List of parsed terms.}

\item{where_specials}{List of term indices by type.}

\item{is_sparse}{Whether data is sparse.}

\item{nyindex}{Number of y-index points.}

\item{nobs}{Number of observations.}

\item{obs_indices}{Observation indices.}

\item{eval_env}{Evaluation environment.}

\item{formula_env}{Formula environment to assign to.}
}
\description{
Expand variables to formula environment for smooth/parametric terms
}
\keyword{internal}
