% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_get_dimensions}
\alias{pffr_get_dimensions}
\title{Detect data dimensions for pffr}
\usage{
pffr_get_dimensions(ydata, data, response_name, frml_env, eval_env)
}
\arguments{
\item{ydata}{The ydata argument (NULL for dense data).}

\item{data}{The data argument.}

\item{response_name}{The response variable name (symbol).}

\item{frml_env}{The formula environment.}

\item{eval_env}{Evaluation environment/list used to resolve variables.}
}
\value{
A list with nobs, nyindex, ntotal, is_sparse.
}
\description{
Detect data dimensions for pffr
}
\keyword{internal}
