% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-robust.R
\name{pffr_gls}
\alias{pffr_gls}
\title{Penalized function-on-function regression with non-i.i.d. residuals
(deprecated)}
\usage{
pffr_gls(
  formula,
  yind,
  hatSigma,
  data = NULL,
  ydata = NULL,
  algorithm = NA,
  method = "REML",
  tensortype = c("ti", "t2"),
  bs.yindex = list(bs = "ps", k = 5, m = c(2, 1)),
  bs.int = list(bs = "ps", k = 20, m = c(2, 1)),
  cond.cutoff = 500,
  ...
)
}
\arguments{
\item{formula, yind, hatSigma, data, ydata, algorithm, method, tensortype, bs.yindex, bs.int, cond.cutoff, ...}{Ignored. The function always errors.}
}
\value{
Does not return; always errors with a deprecation message.
}
\description{
**Deprecated**

`pffr_gls()` is deprecated and errors. Its GLS-based covariance correction
produced poorly calibrated inference in practice. Fit the model with
\code{\link{pffr}()} by REML and use its default intervals (CL2 sandwich with
Satterthwaite critical values; see the section \sQuote{Inference} of
\code{\link{pffr}}).
}
\seealso{
\code{\link{pffr}} for the recommended procedure.
}
\author{
Fabian Scheipl
}
