% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_is_ar1_fit}
\alias{pffr_is_ar1_fit}
\title{Does a fit use an AR(1) working correlation?}
\usage{
pffr_is_ar1_fit(object)
}
\arguments{
\item{object}{A fitted pffr model or its stripped gam version.}
}
\value{
A single logical.
}
\description{
`TRUE` when the fit was estimated by [mgcv::bam()] with a nonzero `rho`
(stored by mgcv as `$AR1.rho`), i.e. on AR(1)-whitened residuals.
}
\keyword{internal}
