% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_is_binary_fit}
\alias{pffr_is_binary_fit}
\title{Does a fit have a binary response?}
\usage{
pffr_is_binary_fit(m)
}
\arguments{
\item{m}{A fitted gam object.}
}
\value{
A single logical.
}
\description{
Does a fit have a binary response?
}
\keyword{internal}
