% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_note_undefined_df}
\alias{pffr_note_undefined_df}
\title{Note (or record) that a working-model moment df was undefined}
\usage{
pffr_note_undefined_df()
}
\value{
`NULL`, invisibly. Called for the side effect.
}
\description{
Note (or record) that a working-model moment df was undefined
}
\keyword{internal}
