% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_process_ff_terms}
\alias{pffr_process_ff_terms}
\title{Process ff/sff terms and assign to formula environment}
\usage{
pffr_process_ff_terms(ff_terms, yind_vec, obs_indices, formula_env)
}
\arguments{
\item{ff_terms}{List of evaluated ff/sff terms.}

\item{yind_vec}{Y-index vector (stacked).}

\item{obs_indices}{Observation indices.}

\item{formula_env}{Formula environment to assign to.}
}
\description{
Process ff/sff terms and assign to formula environment
}
\keyword{internal}
