% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_process_ffpc_terms}
\alias{pffr_process_ffpc_terms}
\title{Process ffpc terms and return formula strings}
\usage{
pffr_process_ffpc_terms(ffpc_terms, obs_indices, yindex_vec_name, formula_env)
}
\arguments{
\item{ffpc_terms}{List of evaluated ffpc terms.}

\item{obs_indices}{Observation indices.}

\item{yindex_vec_name}{Y-index vector name (symbol).}

\item{formula_env}{Formula environment to assign to.}
}
\value{
Character vector of formula strings for ffpc terms.
}
\description{
Process ffpc terms and return formula strings
}
\keyword{internal}
