% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_process_pcre_terms}
\alias{pffr_process_pcre_terms}
\title{Process pcre terms and return formula strings}
\usage{
pffr_process_pcre_terms(
  pcre_terms,
  is_sparse,
  yind,
  yind_vec,
  nyindex,
  nobs,
  obs_indices,
  formula_env
)
}
\arguments{
\item{pcre_terms}{List of evaluated pcre terms.}

\item{is_sparse}{Whether data is sparse.}

\item{yind}{Y-index vector.}

\item{yind_vec}{Y-index vector (stacked).}

\item{nyindex}{Number of y-index points.}

\item{nobs}{Number of observations.}

\item{obs_indices}{Observation indices.}

\item{formula_env}{Formula environment to assign to.}
}
\value{
Character vector of formula strings for pcre terms.
}
\description{
Process pcre terms and return formula strings
}
\keyword{internal}
