% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_setup_response}
\alias{pffr_setup_response}
\title{Setup response and indices for formula environment}
\usage{
pffr_setup_response(
  is_sparse,
  ydata,
  yind,
  yind_name,
  nobs,
  nyindex,
  response_name,
  eval_env,
  frml_env,
  formula_env
)
}
\arguments{
\item{is_sparse}{Whether data is sparse.}

\item{ydata}{The ydata (or NULL).}

\item{yind}{The y-index vector.}

\item{yind_name}{Name of y-index.}

\item{nobs}{Number of observations.}

\item{nyindex}{Number of y-index points.}

\item{response_name}{Response variable name (symbol).}

\item{eval_env}{Evaluation environment.}

\item{frml_env}{Formula environment.}

\item{formula_env}{Environment to assign to.}
}
\value{
A list with yind_vec, yindex_vec_name, obs_indices, missing_indices.
}
\description{
Setup response and indices for formula environment
}
\keyword{internal}
