% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_setup_yind_sparse}
\alias{pffr_setup_yind_sparse}
\title{Setup y-index for sparse data}
\usage{
pffr_setup_yind_sparse(ydata)
}
\arguments{
\item{ydata}{The ydata data.frame.}
}
\value{
A list with yind, yind_name, nyindex.
}
\description{
Setup y-index for sparse data
}
\keyword{internal}
