% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_validate_ydata}
\alias{pffr_validate_ydata}
\title{Validate sparse data format}
\usage{
pffr_validate_ydata(ydata)
}
\arguments{
\item{ydata}{The ydata argument.}
}
\value{
TRUE if valid, stops with error otherwise.
}
\description{
Validate sparse data format
}
\keyword{internal}
