% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.pfr.R
\name{plot.pfr}
\alias{plot.pfr}
\title{Plot a pfr object}
\usage{
\method{plot}{pfr}(x, Qscale = FALSE, ...)
}
\arguments{
\item{x}{a fitted \code{pfr}-object}

\item{Qscale}{For additive functional terms fit with
\code{af(Qtransform=TRUE)}, should the coefficient be plotted on the
raw scale or the quantile-transformed scale?}

\item{...}{arguments handed over to \code{\link[mgcv]{plot.gam}}}
}
\value{
This function's main purpose is its side effect of generating plots.
It also silently returns a list of the data used to produce the plots, which
can be used to generate customized plots.
}
\description{
This function plots the smooth coefficients of a pfr object. These include
functional coefficients as well as any smooths of scalar covariates. The
function simply dispatches to \code{plot.gam}.
}
\author{
Jonathan Gellar
}
\seealso{
\code{\link{af}}, \code{\link{pfr}}
}

