% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dt_basis.R
\name{Predict.matrix.dt.smooth}
\alias{Predict.matrix.dt.smooth}
\title{Predict.matrix method for dt basis}
\usage{
\method{Predict.matrix}{dt.smooth}(object, data)
}
\arguments{
\item{object}{a \code{dt.smooth} object created by
\code{\link{smooth.construct.dt.smooth.spec}}, see
\code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
design matrix for domain-transformed terms
}
\description{
Predict.matrix method for dt basis
}
\author{
Jonathan Gellar
}

