% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpc.R
\name{Predict.matrix.fpc.smooth}
\alias{Predict.matrix.fpc.smooth}
\title{mgcv-style constructor for prediction of FPC terms}
\usage{
\method{Predict.matrix}{fpc.smooth}(object, data)
}
\arguments{
\item{object}{a \code{fpc.smooth} object created by
\code{{smooth.construct.fpc.smooth.spec}}, see
\code{[mgcv]{smooth.construct}}}

\item{data}{see \code{[mgcv]{smooth.construct}}}
}
\value{
design matrix for FPC terms
}
\description{
mgcv-style constructor for prediction of FPC terms
}
\author{
Jonathan Gellar
}
