% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smooth.construct.peer.smooth.spec.R
\name{Predict.matrix.peer.smooth}
\alias{Predict.matrix.peer.smooth}
\title{mgcv-style constructor for prediction of PEER terms}
\usage{
\method{Predict.matrix}{peer.smooth}(object, data)
}
\arguments{
\item{object}{a smooth specification object, see
\code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
design matrix for PC-based functional random effects
}
\description{
mgcv-style constructor for prediction of PEER terms
}
\author{
Fabian Scheipl;  adapted from 'Predict.matrix.random.effect' by S.N. Wood.
}

