% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.pffr}
\alias{print.summary.pffr}
\title{Print method for summary of a pffr fit}
\usage{
\method{print}{summary.pffr}(x, digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{a fitted \code{pffr}-object}

\item{digits}{controls number of digits printed in output.}

\item{signif.stars}{Should significance stars be printed alongside output?}

\item{...}{not used}
}
\value{
A \code{\link{summary.pffr}} object
}
\description{
Pretty printing for a \code{summary.pffr}-object.
See \code{\link[mgcv]{print.summary.gam}()} for details.
}
\author{
Fabian Scheipl, adapted from \code{\link[mgcv]{print.summary.gam}()} by Simon Wood, Henric Nilsson
}

