% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quadWeights.R
\name{quadWeights}
\alias{quadWeights}
\title{Compute quadrature weights}
\usage{
quadWeights(argvals, method = "trapezoidal")
}
\arguments{
\item{argvals}{function arguments.}

\item{method}{quadrature method. Can be either \code{trapedoidal} or \code{midpoint}.}
}
\value{
a vector of quadrature weights for the points supplied in \code{argvals}.
}
\description{
Utility function for numerical integration.
}
\author{
Clara Happ, with modifications by Philip Reiss
}

