% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{re}
\alias{re}
\title{Random effects constructor for fgam}
\usage{
re(x, ...)
}
\arguments{
\item{x}{a grouping variable: must be a \code{factor}}

\item{...}{further arguments handed over to \code{\link[mgcv]{s}}, 
see \code{\link[mgcv]{random.effects}}}
}
\value{
A list with components \code{call} (a rewritten call to
  \code{\link[mgcv]{s}} with \code{bs = "re"}) and \code{data} (a named
  list containing the grouping factor).
}
\description{
Sets up a random effect for the levels of \code{x}. 
Use the \code{by}-argument to request random slopes.
}
\details{
See \code{\link[mgcv]{random.effects}} in \pkg{mgcv}.
}
\seealso{
\code{\link[mgcv]{random.effects}}
}
