\name{refund-package}
\Rdversion{1.1}
\alias{refund-package}
\alias{refund}
\docType{package}
\title{
Regression with Functional Data
}
\description{
Functions for regression with functional data.  The methods currently implemented regress (i) scalar responses on functional predictors (\code{\link{pfr}} [Goldsmith et al., 2011], its longitudinal extension \code{\link{lpfr}} [Goldsmith et al., 2010], and \code{\link{fpcr}} [Reiss and Ogden, 2007]); (ii) functional responses on scalar predictors (\code{\link{fosr}} [Reiss et al., 2010]); and (iii) functional responses on functional predictors (\code{\link{pffr}} [Ivanescu et al., 2011]).}
\details{For a complete list of functions type \code{library(help=refund)}.
}

\author{
Coordinating authors:

Ciprian Crainiceanu \email{ccrainic@jhsph.edu}

Philip Reiss \email{phil.reiss@nyumc.org}


Contributors:

Jeff Goldsmith \email{jgoldsmi@jhsph.edu}

Sonja Greven \email{sonja.greven@stat.uni-muenchen.de}

Lei Huang \email{huangracer@gmail.com}

Fabian Scheipl \email{fabian.scheipl@stat.uni-muenchen.de}


Maintainer: Lei Huang \email{huangracer@gmail.com}
}
\references{
Goldsmith, J., Bobb, J., Crainiceanu, C., Caffo, B., and Reich, D. (2011). Penalized functional regression. \emph{Journal of Computational and Graphical Statistics}, to appear.

Goldsmith, J., Crainiceanu, C., Caffo, B., and Reich, D. (2012). Longitudinal penalized functional regression for cognitive outcomes on neuronal tract measurements. \emph{Journal of the Royal Statistical Society: Series C}, to appear.

Ivanescu, A. E., Staicu, A.-M., Greven, S., Scheipl, F., and Crainiceanu, C. M. (2011).  Penalized function-on-function regression.  Submitted.

Reiss, P. T., Huang, L., and Mennes, M. (2010).  Fast function-on-scalar regression with penalized basis expansions.  \emph{International Journal of Biostatistics}, 6(1), article 28.  Available at \url{http://works.bepress.com/phil_reiss/16/}

Reiss, P. T., and Ogden, R. T. (2007).  Functional principal component regression and functional partial least squares.  \emph{Journal of the American Statistical Association}, 102, 984--996.

}
\keyword{ package }
\examples{
}
