% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{sample_pspline_coef_2d}
\alias{sample_pspline_coef_2d}
\title{Sample coefficients from 2D tensor P-spline prior}
\usage{
sample_pspline_coef_2d(k_s, k_t, wiggliness)
}
\arguments{
\item{k_s}{Basis dimension for first margin.}

\item{k_t}{Basis dimension for second margin.}

\item{wiggliness}{Controls smoothness.}
}
\value{
Matrix of coefficients (k_s x k_t).
}
\description{
Samples from tensor product P-spline prior with additive penalties:
\eqn{P = (1/\mathrm{wiggliness}) \cdot (P_s \otimes I_t + I_s \otimes P_t) + I}{P = (1/wiggliness) * (P_s x I_t + I_s x P_t) + I}
}
\keyword{internal}
