% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fpc.R
\name{smooth.construct.fpc.smooth.spec}
\alias{smooth.construct.fpc.smooth.spec}
\title{Basis constructor for FPC terms}
\usage{
\method{smooth.construct}{fpc.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a \code{fpc.smooth.spec} object, usually generated by a
term \code{s(x, bs="fpc")}; see Details.}

\item{data}{a list containing the data (including any \code{by} variable)
required by this term, with names corresponding to \code{object$term}
(and \code{object$by}). Only the first element of this list is used.}

\item{knots}{not used, but required by the generic \code{smooth.construct}.}
}
\value{
An object of class \code{"fpc.smooth"}. In addtional to the elements
  listed in \code{\link{smooth.construct}}, the object will contain
  \item{sm}{the smooth that is fit in order to generate the basis matrix
    over \code{object$term}}
  \item{V.A}{the matrix of principal components}
}
\description{
Basis constructor for FPC terms
}
\details{
\code{object} must contain an \code{xt} element. This is a list that can
  contain the following elements:
\describe{
  \item{X}{(required) matrix of functional predictors}
  \item{A}{(optional) the number of PC's to retain}
  \item{pve}{(only needed if \code{A} not supplied) the percent variance
    explained used to determine \code{A}}
  \item{bs}{the basis class used to pre-smooth \code{X}; default is \code{"ps"}}
}

Any additional options for the pre-smoothing basis (e.g. \code{k}, \code{m},
  etc.) can be supplied in the corresponding elements of \code{object}.
  See \code{\link[mgcv]{s}} for a full list of options.
}
\author{
Jonathan Gellar \email{JGellar@mathematica-mpr.com}
}
\seealso{
\code{\link{fpcr}}
}

