% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pffr-pcre.R
\name{smooth.construct.pcre.smooth.spec}
\alias{smooth.construct.pcre.smooth.spec}
\title{mgcv-style constructor for PC-basis functional random effects}
\usage{
\method{smooth.construct}{pcre.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, see \code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}

\item{knots}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
An object of class \code{"random.effect"}. See \code{\link[mgcv]{smooth.construct}}
 for the elements that this object will contain.
}
\description{
Sets up design matrix for functional random effects based on the PC scores
of the covariance operator of the random effect process.
See \code{\link[mgcv]{smooth.construct.re.smooth.spec}} for more details on \code{mgcv}-style smoother specification
and \code{\link{pcre}} for the corresponding \code{pffr()}-formula wrapper.
}
\author{
Fabian Scheipl;  adapted from 're' constructor by S.N. Wood.
}

