% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{sym_inv_sqrt}
\alias{sym_inv_sqrt}
\title{Symmetric matrix inverse square root with eigenvalue floor}
\usage{
sym_inv_sqrt(M, tol = 1e-08)
}
\arguments{
\item{M}{Symmetric matrix.}

\item{tol}{Eigenvalue floor for numerical stability.}
}
\value{
Matrix inverse square root of `M`.
}
\description{
Symmetric matrix inverse square root with eigenvalue floor
}
\keyword{internal}
