% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-formula.R
\name{transform_c_term}
\alias{transform_c_term}
\title{Transform c() wrapped terms by removing the wrapper}
\usage{
transform_c_term(term_string)
}
\arguments{
\item{term_string}{Original term string with c() wrapper.}
}
\value{
Term string with c() removed.
}
\description{
Transform c() wrapped terms by removing the wrapper
}
\keyword{internal}
