% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-formula.R
\name{transform_par_term}
\alias{transform_par_term}
\title{Transform parametric terms to varying coefficient terms}
\usage{
transform_par_term(term, yind_vec_name, bs_yindex)
}
\arguments{
\item{term}{The term (call or symbol).}

\item{yind_vec_name}{Name of y-index vector (character).}

\item{bs_yindex}{Basis spec for y-index (named list).}
}
\value{
Transformed term string.
}
\description{
Converts scalar covariates to smooth varying coefficient terms.
}
\keyword{internal}
