% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-formula.R
\name{transform_smooth_term}
\alias{transform_smooth_term}
\title{Transform smooth terms (s, te, ti, t2) to include y-index}
\usage{
transform_smooth_term(term, yindvecname, bs_yindex, tensortype, algorithm)
}
\arguments{
\item{term}{The term call object.}

\item{yindvecname}{Name of y-index vector (character).}

\item{bs_yindex}{Default basis spec for y-index (named list).}

\item{tensortype}{Type of tensor product (symbol: ti or t2).}

\item{algorithm}{Fitting algorithm name (character or symbol).}
}
\value{
Transformed term string.
}
\description{
Converts univariate/bivariate smooth terms into tensor products that
include the functional response index.
}
\keyword{internal}
