% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{validate_and_expand_xind}
\alias{validate_and_expand_xind}
\title{Validate and expand xind to matrix form}
\usage{
validate_and_expand_xind(xind, n, nxgrid, arg_name = "xind")
}
\arguments{
\item{xind}{Vector or matrix of x-index values.}

\item{n}{Number of observations.}

\item{nxgrid}{Number of grid points (columns expected).}

\item{arg_name}{Name of the argument for error messages.}
}
\value{
Matrix of x-index values (n x nxgrid).
}
\description{
Validates that xind has the correct dimensions and is in increasing order,
then expands a vector or single-row matrix to a full n x nxgrid matrix.
Used by [ff()] and [sff()].
}
\keyword{internal}
