% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-formula.R
\name{build_mgcv_data}
\alias{build_mgcv_data}
\title{Build the mgcv data frame from pffr components}
\usage{
build_mgcv_data(newfrmlenv)
}
\arguments{
\item{newfrmlenv}{Environment containing the transformed variables.}
}
\value{
A data frame for mgcv fitting.
}
\description{
Build the mgcv data frame from pffr components
}
\keyword{internal}
