% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CD4-data.R
\docType{data}
\name{cd4}
\alias{cd4}
\title{Observed CD4 cell counts}
\format{A data frame made up of \describe{ \item{list("cd4")}{A 366 x 61
matrix of CD4 cell counts.}\item{ }{A 366 x 61 matrix of CD4 cell counts.}
}}
\description{
CD4 cell counts for 366 subjects between months -18 and 42 since
seroconversion. Each subject's observations are contained in a single row.
}
\references{
Goldsmith, J., Greven, S., and Crainiceanu, C. (2013).
Corrected confidence bands for functional data using principal components.
\emph{Biometrics}, 69(1), 41--51.
}
\keyword{datasets}

