% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.pfr.R
\name{coefficients.pfr}
\alias{coef.pfr}
\alias{coefficients.pfr}
\title{Extract coefficient functions from a fitted pfr-object}
\usage{
\method{coefficients}{pfr}(object, which, n = 100, n2 = 40, se = TRUE,
  seWithMean = TRUE, exclude = FALSE, ...)

\method{coef}{pfr}(object, which, n = 100, n2 = 40, se = TRUE,
  seWithMean = TRUE, exclude = FALSE, ...)
}
\arguments{
\item{object}{return object from \code{\link{pfr}}}

\item{which}{which terms to extract coefficients from (integer vector), defaults
to all terms}

\item{n}{see \code{\link[mgcv]{plot.gam}}}

\item{n2}{see \code{\link[mgcv]{plot.gam}}}

\item{se}{compute standard errors? defaults to TRUE. See \code{\link[mgcv]{plot.gam}}}

\item{seWithMean}{see \code{\link[mgcv]{plot.gam}}}

\item{...}{further arguments passed on to \pkg{mgcv}'s \code{plot.mgcv.smooth}}
}
\value{
a list of data.frames containing the evaluation points,
   coefficient function values and optionally their se's for each term in \code{which}
}
\description{
Extract coefficient functions from a fitted pfr-object
}
\examples{
#TODO: see pfr
}
\author{
Fabian Scheipl, se-computation parts adapted from Simon Wood's \code{plot.gam}.
}

