% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{compute_coef_se}
\alias{compute_coef_se}
\title{Compute standard errors for coefficient extraction}
\usage{
compute_coef_se(X, trmind, trm, object_info, covmat, seWithMean)
}
\arguments{
\item{X}{Prediction matrix.}

\item{trmind}{Index vector for term parameters.}

\item{trm}{Smooth term object (for nCons and meanL1).}

\item{object_info}{List with cmX, Vp.}

\item{covmat}{Covariance matrix.}

\item{seWithMean}{Logical, use seWithMean approach?}
}
\value{
Numeric vector of standard errors.
}
\description{
Compute standard errors for coefficient extraction
}
\keyword{internal}
