% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{draw_coef_perturbations}
\alias{draw_coef_perturbations}
\title{Draw coefficient perturbations for simultaneous intervals}
\usage{
draw_coef_perturbations(covmat, n_sim, sim_seed = NULL)
}
\arguments{
\item{covmat}{Covariance matrix.}

\item{n_sim}{Number of simulation draws.}

\item{sim_seed}{Optional integer seed.}
}
\value{
Matrix with one simulated perturbation vector per column.
}
\description{
Draw coefficient perturbations for simultaneous intervals
}
\keyword{internal}
