% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/f_sum2.R
\name{f_sum2}
\alias{f_sum2}
\title{Sum computation 2}
\usage{
f_sum2(y, fixef, mu.q.c, kt, theta)
}
\arguments{
\item{y}{}

\item{fixef}{}

\item{mu.q.c}{}

\item{kt}{}

\item{theta}{}
}
\description{
Internal function used compute a sum in FPCA-based covariance updates
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

