% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sum4.R
\name{f_sum4}
\alias{f_sum4}
\title{Sum computation 2}
\usage{
f_sum4(mu.q.c, sig.q.c, mu.q.bpsi, sig.q.bpsi, theta, obspts.mat)
}
\arguments{
\item{mu.q.c}{current value of mu.q.c}

\item{sig.q.c}{current value of sig.q.c}

\item{mu.q.bpsi}{current value of mu.q.bpsi}

\item{sig.q.bpsi}{current value of sig.q.bpsi}

\item{theta}{current value of theta}

\item{obspts.mat}{matrix indicating where curves are observed}
}
\description{
Internal function used compute a sum in FPCA-based covariance updates
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

