% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_trace.R
\name{f_trace}
\alias{f_trace}
\title{Trace computation}
\usage{
f_trace(Theta_i, Sig_q_Bpsi, Kp, Kt)
}
\arguments{
\item{Theta_i}{basis functions on observed grid points}

\item{Sig_q_Bpsi}{variance of FPC basis coefficients}

\item{Kp}{number of FPCs}

\item{Kt}{number of spline basis functions}
}
\description{
Internal function used compute a trace in FPCA-based covariance updates
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

