% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\docType{data}
\name{ff_presets}
\alias{ff_presets}
\title{Function-on-function effect presets}
\format{
An object of class \code{list} of length 5.
}
\usage{
ff_presets
}
\description{
Named list of preset beta(s,t) coefficient surfaces for ff() terms.
Each function takes arguments (s, t) and returns a matrix.
}
\keyword{internal}
