% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-core.R
\name{pffr_build_metadata}
\alias{pffr_build_metadata}
\title{Build the pffr metadata list}
\usage{
pffr_build_metadata(
  call,
  formula,
  term_map,
  label_map,
  short_labels,
  response_name,
  nobs,
  nyindex,
  yind_name,
  yind,
  where_specials,
  ff_terms,
  ffpc_terms,
  pcre_terms,
  missing_indices,
  is_sparse,
  ydata,
  sandwich,
  dof_correction = "none",
  edf_type = "trace"
)
}
\description{
Build the pffr metadata list
}
\keyword{internal}
