% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-robust.R
\name{pffrGLS}
\alias{pffrGLS}
\title{Penalized function-on-function regression with non-i.i.d. residuals
(deprecated)}
\usage{
pffrGLS(
  formula,
  yind,
  hatSigma,
  data = NULL,
  ydata = NULL,
  algorithm = NA,
  method = "REML",
  tensortype = c("ti", "t2"),
  bs.yindex = list(bs = "ps", k = 5, m = c(2, 1)),
  bs.int = list(bs = "ps", k = 20, m = c(2, 1)),
  cond.cutoff = 500,
  ...
)
}
\arguments{
\item{formula, yind, hatSigma, data, ydata, algorithm, method, tensortype, bs.yindex, bs.int, cond.cutoff, ...}{Ignored. The function always errors.}
}
\value{
Does not return; always errors with a deprecation message.
}
\description{
**Deprecated**

`pffrGLS()` is deprecated. Use \code{\link{pffr}()} with
\code{sandwich = "cluster"} or \code{sandwich = "cl2"} instead.
}
\keyword{internal}
