% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pfr.R
\name{pfr}
\alias{pfr}
\title{Penalized Functional Regression}
\usage{
pfr(formula = NULL, fitter = NA, ...)
}
\arguments{
\item{formula}{a formula with special terms as for \code{\link{gam}},
with additional special terms \code{\link{af}}() and \code{\link{lf}}().}

\item{fitter}{the name of the function used to estimate the model. Defaults
to \code{\link{gam}} if the matrix of functional responses has less than 2e5
data points and to \code{\link{bam}} if not. "gamm" (see \code{\link{gamm}})
and "gamm4" (see \code{\link{gamm4}}) are valid options as well.}

\item{...}{additional arguments that are valid for \code{\link{gam}} or
\code{\link{bam}}; for example, specify a \code{gamma} > 1 to increase amount
of smoothing when using GCV to choose smoothing parameters or
\code{method="REML"} to change to REML for estimation of smoothing parameters
(default is GCV).}
}
\value{
A fitted pfr-object, which is a \code{\link{gam}}-object with some
additional information in a pfr-element If fitter is "gamm" or "gamm4", only
the $gam part of the returned list is modified in this way.
}
\description{
Implements various approaches to penalized scalar-on-function regression.
These techniques include Penalized Functional Regression (Goldsmith et al.,
2011), Functional Principal Component Regression (Reiss and Ogden, 2007),
Functional Generalized Additive Models (McLean et al., 2013), and Partially
Empirical Eigenvectors for Regression (Randolph et al., 2012).
This function is a wrapper for mgcv's \code{\link{gam}}
and its siblings to fit models with a scalar (but necessarily continuous)
response.
}
\section{Warning}{

Binomial responses should be specified as a numeric vector rather than as a
matrix or a factor.
}
\author{
Mathew W. McLean \email{mathew.w.mclean@gmail.com} and Fabian Scheipl
}
\references{
McLean, M. W., Hooker, G., Staicu, A.-M., Scheipl, F., and
Ruppert, D. (2014). Functional generalized additive models. \emph{Journal of
Computational and Graphical Statistics}, \bold{23 (1)}, pp. 249-269.
Available at \url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3982924}.
}
\seealso{
\code{\link{af}}, \code{\link{lf}}
}

