% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{plot.pffr}
\alias{plot.pffr}
\title{Plot a pffr fit}
\usage{
\method{plot}{pffr}(x, ...)
}
\arguments{
\item{x}{a fitted \code{pffr}-object}

\item{...}{arguments handed over to \code{\link[mgcv]{plot.gam}}}
}
\value{
This function only generates plots.
}
\description{
Plot a fitted pffr-object. Simply dispatches to \code{\link[mgcv]{plot.gam}}.
}
\author{
Fabian Scheipl
}
