% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.R
\name{Predict.matrix.peer.smooth}
\alias{Predict.matrix.peer.smooth}
\title{mgcv-style constructor for prediction of PEER terms}
\usage{
\method{Predict.matrix}{peer.smooth}(object, data)
}
\arguments{
\item{object}{a \code{peer.smooth} object created by
\code{\link{smooth.construct.peer.smooth.spec}}, see
\code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
design matrix for PEER terms
}
\description{
mgcv-style constructor for prediction of PEER terms
}
\author{
Jonathan Gellar
}
