\name{refund-package}
\Rdversion{1.1}
\alias{refund-package}
\alias{refund}
\docType{package}
\title{
Regression with Functional Data
}
\description{
Functions for regression with functional data.  The principal methods currently implemented are (1) functional-response/scalar-predictor 
regression by penalized OLS or GLS; and (2) functional principal component regression for scalar responses and functional predictors.}
\details{For a complete list of functions type \code{library(help=refund)}.
}

\author{
Philip Reiss \email{phil.reiss@nyumc.org} and Lei Huang \email{huangracer@gmail.com}

Maintainer: Lei Huang \email{huangracer@gmail.com}
}
\references{
}
\keyword{ package }
\examples{
## see examples for fosr
}
