\name{refund-package}
\Rdversion{1.1}
\alias{refund-package}
\alias{refund}
\docType{package}
\title{
Regression with Functional Data
}
\description{
Methods for regression with functional data. Various approaches to regression with scalar responses and functional predictors are implemented by \code{\link{pfr}}, \code{\link{peer}}, \code{\link{lpeer}} and \code{\link{fpcr}}. For regression with functional responses, see \code{\link{pffr}}, \code{\link{fosr}}, and \code{\link{fosr2s}}.}
\details{For a complete list of functions type \code{library(help=refund)}.
}

\author{
\emph{Authors:}
Lei Huang \email{huangracer@gmail.com}, Fabian Scheipl \email{fabian.scheipl@stat.uni-muenchen.de}, Jeff Goldsmith \email{jeff.goldsmith@columbia.edu},  Ciprian Crainiceanu \email{ccrainic@jhsph.edu}, Philip Reiss \email{phil.reiss@nyumc.org}.


\emph{Contributors:}
Lan Huo, Bruce Swihart, Sonja Greven, Jaroslaw Harezlak, Madan Gopal Kundu, Yihong Zhao, Mathew W. McLean, Luo Xiao, Jonathan Gellar

\emph{Maintainer:} Lei Huang \email{huangracer@gmail.com}
}
\keyword{ package }

