% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pffr-methods.R
\name{residuals.pffr}
\alias{residuals.pffr}
\title{Obtain residuals for a pffr fit}
\usage{
\method{residuals}{pffr}(object, reformat = TRUE, ...)
}
\arguments{
\item{object}{a fitted \code{pffr}-object}

\item{reformat}{logical, defaults to TRUE. Should residuals be returned in \code{n x yindex} matrix
form (default for regular grid data) or in the shape of the originally supplied \code{ydata}
argument or (default for sparse/irregular data), or, if \code{FALSE}, simply in the
long vector shape returned by \code{resid.gam()}?}

\item{...}{other arguments, passed to \code{\link[mgcv]{residuals.gam}}.}
}
\value{
A matrix or vector of residuals
}
\description{
Obtain residuals for a pffr fit
}
\author{
Fabian Scheipl
}

