% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\name{simplify_term_label}
\alias{simplify_term_label}
\title{Simplify a term label for display}
\usage{
simplify_term_label(term_name, yind_name)
}
\arguments{
\item{term_name}{Original term name string.}

\item{yind_name}{Name of the y-index variable.}
}
\value{
Simplified label string.
}
\description{
Simplify a term label for display
}
\keyword{internal}
