% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffr-utilities.R
\docType{data}
\name{smooth_presets}
\alias{smooth_presets}
\title{Smooth effect presets for s() terms}
\format{
An object of class \code{list} of length 6.
}
\usage{
smooth_presets
}
\description{
Named list of preset f(x,t) surfaces for smooth varying coefficient terms.
Each function takes arguments (x, t) and returns a matrix.
}
\keyword{internal}
